requirements.txt
# Core dependencies
crowdstrike-falconpy>=1.3.0
mcp>=1.4.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0

# Add mcp-use and its dependencies
mcp-use
langchain-openai
